+++ test/acceptance/test.rb
@@ -32,21 +32,11 @@ class AcceptanceTest < Minitest::Test
   end
 
   def test_active_record_load_schema
-    if ActiveRecord::Base.respond_to?(:connection_db_config)
-      ActiveRecord::Base.establish_connection url: uri, use_metadata_table: false
-      db_config = ActiveRecord::Base.connection_db_config
-      options = []
-    else
-      db_config = {
-        "adapter" => "postgresql",
-        "url" => uri,
-        "use_metadata_table" => false
-      }
-      options = ["test"]
-    end
+    ActiveRecord::Base.establish_connection url: uri, use_metadata_table: false
+    db_config = ActiveRecord::Base.connection_db_config
 
     _, stderr = capture_subprocess_io do
-      ActiveRecord::Tasks::DatabaseTasks.load_schema db_config, :sql, File.expand_path("structure.sql", __dir__), *options
+      ActiveRecord::Tasks::DatabaseTasks.load_schema db_config, :sql, File.expand_path("structure.sql", __dir__)
     end
 
     assert_includes stderr, "🚀"