+++ CHANGELOG.md
@@ -6,7 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
-No notable changes.
+### Changed
+* Require Ruby ≥ 3.0, Active Record ≥ 6.1, and PG ≥ 1.1 ([#492](https://github.com/haines/pg-aws_rds_iam/pull/492))
 
 ## [0.4.2] - 2023-01-10